+++ command/container/new.go
@@ -2,7 +2,6 @@ package container
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -29,34 +28,7 @@ func newCommand(home string, docker client.CommonAPIClient, output terminal.Outp
   # expand the number of images from the search
   nitro container new --limit 50`,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
-			// verify the config exists
-			_, err := config.Load(home)
-			if errors.Is(err, config.ErrNoConfigFile) {
-				output.Info("Warning:", err.Error())
-
-				// ask if the init command should run
-				init, err := output.Confirm("Run `nitro init` now to create the config", true, "?")
-				if err != nil {
-					return err
-				}
-
-				// if init is false return nil
-				if !init {
-					return fmt.Errorf("You must run `nitro init` in order to add a site...")
-				}
-
-				// run the init command
-				for _, c := range cmd.Parent().Commands() {
-					// set the init command
-					if c.Use == "init" {
-						if err := c.RunE(c, args); err != nil {
-							return err
-						}
-					}
-				}
-			}
-
-			return nil
+			return prompt.VerifyInit(cmd, args, home, output)
 		},
 		PostRunE: func(cmd *cobra.Command, args []string) error {
 			return prompt.RunApply(cmd, args, false, output)